requirements.txt
requests
python-dotenv
boto3
prefect
prefect-aws